Migrate client.modules to new plugin system
