--- client/modules/Time.py
@@ -1,35 +1,0 @@
-# -*- coding: utf-8-*-
-import datetime
-import re
-from client.app_utils import get_timezone
-from semantic.dates import DateService
-
-WORDS = ["TIME"]
-
-
-def handle(text, mic, profile):
-    """
-        Reports the current time based on the user's timezone.
-
-        Arguments:
-        text -- user-input, typically transcribed speech
-        mic -- used to interact with the user (for both input and output)
-        profile -- contains information related to the user (e.g., phone
-                   number)
-    """
-
-    tz = get_timezone(profile)
-    now = datetime.datetime.now(tz=tz)
-    service = DateService()
-    response = service.convertTime(now)
-    mic.say("It is %s right now." % response)
-
-
-def is_valid(text):
-    """
-        Returns True if input is related to the time.
-
-        Arguments:
-        text -- user-input, typically transcribed speech
-    """
-    return bool(re.search(r'\btime\b', text, re.IGNORECASE))
