+++ plugins/speechhandler/birthday/plugin.info
@@ -1,0 +1,10 @@
+[Plugin]
+Name = birthday
+Version = 1.0.0
+License = MIT
+URL = http://jasperproject.github.io/
+Description = Keep track of Facebook friends' birthdays
+
+[Author]
+Name = Jasper Project
+URL = http://jasperproject.github.io/
