+++ plugins/speechhandler/joke/plugin.info
@@ -1,0 +1,10 @@
+[Plugin]
+Name = joke
+Version = 1.0.0
+License = MIT
+URL = http://jasperproject.github.io/
+Description = Tells you a cheezy "Knock Knock" joke
+
+[Author]
+Name = Jasper Project
+URL = http://jasperproject.github.io/
