+++ plugins/speechhandler/life/life.py
@@ -1,0 +1,34 @@
+# -*- coding: utf-8-*-
+import random
+import re
+from client import plugin
+
+
+class MeaningOfLifePlugin(plugin.SpeechHandlerPlugin):
+    def get_phrases(self):
+        return ["MEANING", "OF", "LIFE"]
+
+    def handle(self, text, mic):
+        """
+        Responds to user-input, typically speech text, by relaying the
+        meaning of life.
+
+        Arguments:
+        text -- user-input, typically transcribed speech
+        mic -- used to interact with the user (for both input and output)
+        """
+        messages = ["It's 42, you idiot.",
+                    "It's 42. How many times do I have to tell you?"]
+
+        message = random.choice(messages)
+
+        mic.say(message)
+
+    def is_valid(self, text):
+        """
+        Returns True if the input is related to the meaning of life.
+
+        Arguments:
+        text -- user-input, typically transcribed speech
+        """
+        return bool(re.search(r'\bmeaning of life\b', text, re.IGNORECASE))
