+++ plugins/speechhandler/news/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8-*-
+from .news import NewsPlugin
