+++ plugins/speechhandler/notifications/plugin.info
@@ -1,0 +1,10 @@
+[Plugin]
+Name = notifications
+Version = 1.0.0
+License = MIT
+URL = http://jasperproject.github.io/
+Description = Get notifications for Facebook events
+
+[Author]
+Name = Jasper Project
+URL = http://jasperproject.github.io/
