+++ plugins/speechhandler/unclear/plugin.info
@@ -1,0 +1,10 @@
+[Plugin]
+Name = unclear
+Version = 1.0.0
+License = MIT
+URL = http://jasperproject.github.io/
+Description = Dummy module to handle unknown speech
+
+[Author]
+Name = Jasper Project
+URL = http://jasperproject.github.io/
