+++ plugins/speechhandler/weather/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8-*-
+from .weather import WeatherPlugin
