+++ plugins/speechhandler/weather/plugin.info
@@ -1,0 +1,10 @@
+[Plugin]
+Name = weather
+Version = 1.0.0
+License = MIT
+URL = http://jasperproject.github.io/
+Description = Get weather forecasts from wunderground.com
+
+[Author]
+Name = Jasper Project
+URL = http://jasperproject.github.io/
